Separated and cleaned up StyleParam
 - consolidated variant types
 - removed duplication of string constants
 - created separate file for StyleParam struct
