+++ core/src/style/polylineStyle.cpp
@@ -34,17 +34,24 @@ void PolylineStyle::constructShaderProgram() {
 PolylineStyle::Parameters PolylineStyle::parseRule(const DrawRule& _rule) const {
     Parameters p;
 
+    uint32_t cap = 0, join = 0;
+
     _rule.get(StyleParamKey::order, p.order);
     _rule.get(StyleParamKey::color, p.color);
     _rule.get(StyleParamKey::width, p.width);
-    _rule.get(StyleParamKey::cap, p.cap);
-    _rule.get(StyleParamKey::join, p.join);
+    _rule.get(StyleParamKey::cap, cap);
+    _rule.get(StyleParamKey::join, join);
+
+    p.cap = static_cast<CapTypes>(cap);
+    p.join = static_cast<JoinTypes>(join);
 
     if (_rule.get(StyleParamKey::outline_color, p.outlineColor) |
         _rule.get(StyleParamKey::outline_width, p.outlineWidth) |
-        _rule.get(StyleParamKey::outline_cap, p.outlineCap) |
-        _rule.get(StyleParamKey::outline_join, p.outlineJoin)) {
+        _rule.get(StyleParamKey::outline_cap, cap) |
+        _rule.get(StyleParamKey::outline_join, join)) {
         p.outlineOn = true;
+        p.outlineCap = static_cast<CapTypes>(cap);
+        p.outlineJoin = static_cast<JoinTypes>(join);
     }
 
     return p;
